Problems encountered in make_netlist() function
The deduction of which nodes to connect for series connections does not work for an arbitrary string representation. This is because it is impossible to tell which direction the cells have to be connected to each other for a series connection. This may mean that a third connection dimension is needed.